# api/openapi.yaml
# Generated with protoc-gen-openapi
# https://github.com/google/gnostic/tree/master/cmd/protoc-gen-openapi

openapi: 3.0.3
info:
    title: ""
    version: 0.0.1
paths:
    /v1/class/create:
        post:
            tags:
                - Class
            description: 创建班级
            operationId: Class_CreateClass
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/http.class.CreateClassRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/http.class.CreateClassReply'
    /v1/class/get_all:
        post:
            tags:
                - Class
            description: 获取所有班级列表
            operationId: Class_GetAllClass
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/http.class.GetAllClassRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/http.class.GetAllClassReply'
    /v1/class/get_by_name:
        post:
            tags:
                - Class
            description: 根据名字获取班级信息
            operationId: Class_GetClassByName
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/http.class.GetClassByNameRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/http.class.GetClassByNameReply'
    /v1/class/join:
        post:
            tags:
                - Class
            description: 加入班级
            operationId: Class_JoinClass
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/http.class.JoinClassRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/http.class.JoinClassReply'
    /v1/class/set_attendance:
        post:
            tags:
                - Class
            description: 设置考勤
            operationId: Class_SetAttendance
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/http.class.SetAttendanceRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/http.class.SetAttendanceReply'
    /v1/class/student_sign_in:
        post:
            tags:
                - Class
            description: 学生签到
            operationId: Class_StudentSignIn
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/http.class.StudentSignInRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/http.class.StudentSignInReply'
    /v1/user/getuser:
        post:
            tags:
                - User
            operationId: User_GetCurrentUser
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/http.user.GetCurrentUserRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/http.user.GetCurrentUserReply'
    /v1/user/list:
        post:
            tags:
                - User
            operationId: User_List
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/http.user.ListRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/http.user.ListReply'
    /v1/user/listbypage:
        post:
            tags:
                - User
            operationId: User_ListUserByPage
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/http.user.ListUserByPageRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/http.user.ListUserByPageReply'
    /v1/user/login:
        post:
            tags:
                - User
            description: 用户端登录
            operationId: User_Login
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/http.user.LoginRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/http.user.LoginReply'
    /v1/user/register:
        post:
            tags:
                - User
            description: 用户注册
            operationId: User_Register
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/http.user.RegisterRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/http.user.RegisterReply'
components:
    schemas:
        http.class.CreateClassReply:
            type: object
            properties:
                id:
                    type: string
        http.class.CreateClassRequest:
            type: object
            properties:
                name:
                    type: string
            description: 创建班级
        http.class.GetAllClassReply:
            type: object
            properties:
                name:
                    type: array
                    items:
                        type: string
        http.class.GetAllClassRequest:
            type: object
            properties: {}
            description: 获取所有班级列表
        http.class.GetClassByNameReply:
            type: object
            properties:
                name:
                    type: string
                student:
                    type: string
                teacher:
                    type: string
                studentAbsence:
                    type: string
                studentAbsenceCount:
                    type: string
        http.class.GetClassByNameRequest:
            type: object
            properties:
                name:
                    type: string
            description: 获取班级信息
        http.class.JoinClassReply:
            type: object
            properties:
                id:
                    type: string
        http.class.JoinClassRequest:
            type: object
            properties:
                name:
                    type: string
            description: 加入班级
        http.class.SetAttendanceReply:
            type: object
            properties:
                id:
                    type: string
        http.class.SetAttendanceRequest:
            type: object
            properties:
                time:
                    type: integer
                    format: int32
                attendName:
                    type: string
            description: 设置考勤
        http.class.StudentSignInReply:
            type: object
            properties:
                message:
                    type: string
        http.class.StudentSignInRequest:
            type: object
            properties:
                attendName:
                    type: string
            description: 学生签到
        http.user.GetCurrentUserReply:
            type: object
            properties:
                id:
                    type: string
                account:
                    type: string
                password:
                    type: string
                username:
                    type: string
                avatarUrl:
                    type: string
                gender:
                    type: integer
                    format: int32
                userInfo:
                    type: string
                userRole:
                    type: integer
                    format: int32
        http.user.GetCurrentUserRequest:
            type: object
            properties: {}
        http.user.ListReply:
            type: object
            properties:
                username:
                    type: string
                avatarUrl:
                    type: string
                userInfo:
                    type: string
        http.user.ListRequest:
            type: object
            properties:
                id:
                    type: string
        http.user.ListUserByPageReply:
            type: object
            properties:
                userList:
                    type: array
                    items:
                        $ref: '#/components/schemas/http.user.UserInfo'
                total:
                    type: integer
                    format: int32
        http.user.ListUserByPageRequest:
            type: object
            properties:
                current:
                    type: integer
                    format: int32
                pageSize:
                    type: integer
                    format: int32
        http.user.LoginReply:
            type: object
            properties:
                id:
                    type: string
                token:
                    type: string
            description: 返回前端的请求
        http.user.LoginRequest:
            type: object
            properties:
                account:
                    type: string
                password:
                    type: string
            description: |-
                Register 用户登录
                 前端传来的请求
        http.user.RegisterReply:
            type: object
            properties:
                id:
                    type: string
            description: 返回前端的请求
        http.user.RegisterRequest:
            type: object
            properties:
                account:
                    type: string
                password:
                    type: string
                checkPassword:
                    type: string
            description: |-
                Register 用户注册
                 前端传来的请求
        http.user.UserInfo:
            type: object
            properties:
                id:
                    type: string
                account:
                    type: string
                password:
                    type: string
                username:
                    type: string
                avatarUrl:
                    type: string
                gender:
                    type: integer
                    format: int32
                userInfo:
                    type: string
                userRole:
                    type: integer
                    format: int32
tags:
    - name: Class
    - name: User